requirements.txt
Flask==1.1.2
PyNaCl==1.4.0
requests==2.24.0
